allow cross compilation, partially migrate to new syntax

+++ core/src/test/scala/com/avsystem/commons/serialization/cbor/HFloatTest.scala
@@ -9,7 +9,7 @@ class HFloatTest extends AnyFunSuite {
     val float = java.lang.Float.intBitsToFloat(floatBits)
     test(name) {
       val hfloat = HFloat.fromFloat(float)
-      assert(hfloat.raw.toHexString == hfloatBits.toShort.toHexString)
+      assert(hfloat.raw.toInt.toHexString == hfloatBits.toInt.toHexString)
       assert(java.lang.Float.floatToIntBits(hfloat.toFloat).toHexString == roundedBits.toHexString)
     }
   }
